cast numpy floats to float in doctests

+++ sknetwork/classification/pagerank.py
@@ -51,7 +51,7 @@ class PageRankClassifier(RankClassifier):
     >>> labels_true = graph.labels
     >>> labels = {0: labels_true[0], 33: labels_true[33]}
     >>> labels_pred = pagerank.fit_predict(adjacency, labels)
-    >>> np.round(np.mean(labels_pred == labels_true), 2)
+    >>> float(np.round(np.mean(labels_pred == labels_true), 2))
     0.97
 
     References
